Prepare a seperate Op for zero-width check
This allows us to avoid making opMatchContext for every iteration,
by making a context once at the begining of repeating.
However, this is slightly slower than the current implementation,
maybe because the overhead of Op dispatching since this increases
a branch in the switch statement of dispatching, that is executed
every time an Op is dispatched.